+++ src/hoist/_schema.py
@@ -11,8 +11,19 @@ __all__ = (
 )
 
 
-def verify_schema(schema: Schema, data: Payload) -> None:
+def _update_schema(schema: Schema) -> Schema:
+    res: Schema = {}
+
+    for k, v in schema.items():
+        args = getattr(v, "__args__", None)
+        res[k] = v if not args else args
+
+    return res
+
+
+def verify_schema(raw_schema: Schema, data: Payload) -> None:
     """Verify that a payload matches the schema."""
+    schema = _update_schema(raw_schema)
     for key, typ in schema.items():
         value = data.get(key)
         if key is Any:
@@ -20,7 +31,7 @@ def verify_schema(schema: Schema, data: Payload) -> None:
 
         vtype = type(value) if value is not None else None
 
-        if isinstance(typ, tuple):
+        if isinstance(typ, (tuple, list)):
             if vtype not in typ:
                 log(
                     "schema validation",